Create SinglyLinkedList
I have created a SinglyLinkedList implemented manually. I have also created the class SinglyLinkedListNode that will be one unit from the Singly Linked List. The list will be generic. I have created the class "Books" only as an example of Object that can be added to the list. 
The implementation has also a main class. In the main class i have tried to communicate with the user through some messages.